requirements.txt
paho-mqtt==1.6.1
python-dotenv==1.0.1
requests==2.3
python-snap7==1.3
pymodbus==2.4.3